Refactor TriblerConfig
